requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2

# LangGraph / LangChain
langgraph==0.2.35
langchain==0.3.4
langchain-google-genai==2.0.5
langchain-community==0.3.4
langchain-core==0.3.12
google-generativeai==0.8.3

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
redis[hiredis]==5.1.1
psycopg2-binary==2.9.9

# Vector DB
chromadb==0.5.15
sentence-transformers==3.2.1

# Security / Threat Intel
python-nmap==0.7.1
requests==2.32.3
aiohttp==3.10.10
ipwhois==1.3.0

# Log Processing
elasticsearch==8.15.1
structlog==24.4.0

# ML / Analysis
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
joblib==1.4.2

# Monitoring
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Utilities
celery==5.4.0
apscheduler==3.10.4
click==8.1.7
rich==13.9.2
typer==0.12.5
jinja2==3.1.4
